refactor(MatchingCountResult): MatchingCountResult을 model로 변경
