+++ src/main/java/lotto/dto/LottoResult.java
@@ -1,6 +1,7 @@
 package lotto.dto;
 
 import java.util.List;
+import lotto.model.MatchingCountResult;
 
 public record LottoResult(List<MatchingCountResult> matchingCountResults, double rate) {
 }
