+++ src/main/java/lotto/model/MatchingCountResult.java
@@ -1,4 +1,4 @@
-package lotto.dto;
+package lotto.model;
 
 import lotto.constant.WinningCondition;
